+++ apps/docs/app/components/common/Category.tsx
@@ -1,3 +1,9 @@
+import PersonIcon from '@mui/icons-material/Person';
+
+import { CategoryStyle } from './styles';
+
+//style
+
 import styled, { css } from 'styled-components';
 
 interface StyleProps {
@@ -29,3 +35,20 @@ export const CategoryStyle = styled.div<StyleProps>`
     font-size: 18px;
   }
 `;
+
+
+interface Props {
+  title: string;
+  active: boolean;
+}
+
+const Category = ({ title, active }: Props) => {
+  return (
+    <CategoryStyle active={active}>
+      <PersonIcon className='icon' />
+      <span>{title}</span>
+    </CategoryStyle>
+  );
+};
+
+export default Category;
