+++ apps/docs/tsconfig.json
@@ -13,6 +13,6 @@
     "next-env.d.ts",
     "next.config.ts",
     ".next/types/**/*.ts"
-  ],
+, "../../packages/ui/src/components/common/Badge.tsx"  ],
   "exclude": ["node_modules"]
 }
