+++ packages/ui/src/components/common/Avatar.tsx
@@ -1,3 +1,18 @@
+import { ProfileImgStyle } from './styles';
+import { ImgSizeKey, Variant } from './type';
+
+export type ImgSizeKey = 'xs' | 'sm' | 'md' | 'lg' | 'xl' | 'xxl';
+export type Variant = 'circle' | 'rounded';
+
+interface Props {
+  size: ImgSizeKey;
+  variant: Variant;
+  imgUrl: string;
+  alt: string;
+  className?: string;
+}
+
+//style
 import styled from 'styled-components';
 
 import { ImgSizeKey, Variant } from './type';
@@ -35,3 +50,17 @@ export const ProfileImgStyle = styled.div.attrs<ImgProps>(
     };
   },
 )<ImgProps>``;
+
+
+const ProfileImg = ({ size, variant, imgUrl, alt }: Props) => {
+  return (
+    <ProfileImgStyle
+      size={size}
+      variant={variant}
+      imgUrl={imgUrl}
+      className='m-2'
+    />
+  );
+};
+
+export default ProfileImg;
