--- packages/ui/src/components/common/Button/index.tsx
@@ -1,30 +1,0 @@
-'use client';
-
-import { ReactNode } from 'react';
-
-import { ButtonStyle } from './styles';
-
-export interface Props {
-  size: string;
-  color: string;
-  variant: string;
-  onClick: () => void;
-  children: ReactNode;
-}
-
-const Button = ({ size, color, variant, children, onClick }: Props) => {
-  const disabled = color === 'disabled' ? true : false;
-  return (
-    <ButtonStyle
-      size={size}
-      color={color}
-      variant={variant}
-      onClick={onClick}
-      disabled={disabled}
-    >
-      {children}
-    </ButtonStyle>
-  );
-};
-
-export default Button;
