--- packages/ui/src/components/common/Profile/ProfileImg/index.tsx
@@ -1,23 +1,0 @@
-import { ProfileImgStyle } from './styles';
-import { ImgSizeKey, Variant } from './type';
-
-interface Props {
-  size: ImgSizeKey;
-  variant: Variant;
-  imgUrl: string;
-  alt: string;
-  className?: string;
-}
-
-const ProfileImg = ({ size, variant, imgUrl, alt }: Props) => {
-  return (
-    <ProfileImgStyle
-      size={size}
-      variant={variant}
-      imgUrl={imgUrl}
-      className='m-2'
-    />
-  );
-};
-
-export default ProfileImg;
