--- packages/ui/src/components/common/Profile/ProfileImg/type.d.ts
@@ -1,2 +1,0 @@
-export type ImgSizeKey = 'xs' | 'sm' | 'md' | 'lg' | 'xl' | 'xxl';
-export type Variant = 'circle' | 'rounded';
